+++ CHANGELOG.md
@@ -34,3 +34,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 ## [1.0.5] - 2020-08-24
 ### Added
 - Bugfix : Capture CPU does not account CRITICAL AND WARNING CPU data
+
+## [1.0.6] - 2020-09-01
+### Added
+- Feature : Added support to sync environment with on the fly masking jobs
+- Bugfix  : Provide user friendly message if job does not exists in any engine 
+- BufFix  : Set cpu as 0 if not able to connect VE mgmt stack and proceed